+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 01. :gift: Added class `dev.orne.http.client.AbstractHttpServiceOperation`
 01. :gift: Added class `dev.orne.http.client.AbstractStatusIndependentOperation`
 01. :gift: Added class `dev.orne.http.client.AbstractStatusIndependentGetOperation`
+01. :gift: Added class `dev.orne.http.client.AbstractStatusIndependentPutOperation`
 01. :gift: Added class `dev.orne.http.client.AbstractStatusDependentOperation`
 01. :gift: Added class `dev.orne.http.client.AbstractStatusDependentGetOperation`
 01. :gift: Added class `dev.orne.http.client.VoidResponseHandler`